Updated to the latest code.
And below a list of comments:
8/12/2020 Scaling Factor is calculated before setting the Width and Height of the Rectangle
1/11/2016 Minor DocumentImage refactoring
21/7/2016 Avoid errors in FileManager when running PASSPORT, due to building DocumentImage in Release x64
3/7/2015 Fixes extra white pixel bug on AnswerSheet's rightmost/bottom pixels and optimizes iPhone photo processing

+++ Sources/Math/Geometry/ClosePointsMergingOptimizer.cs
@@ -94,7 +94,7 @@ namespace AForge.Math.Geometry
             }
             else
             {
-                float distance = 0;
+                double distance = 0;
 
                 // add first point to the new shape
                 optimizedShape.Add( shape[0] );
